src/currencyservice/requirements.txt
grpcio 
grpcio-tools 
protobuf 
googleapis-common-protos
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-grpc
opentelemetry-exporter-jaeger
